bres: experiments in triangle land
I need to assemble scanlines to send to newport to render as spans.

To do this I will need to be able to assemble a list of scanlines to
render to represent a triangle, and it starts here.

This is the shared start point start, flat bottom finish triangle.
Next is the flat top, finish at point triangle.

Then I can join them together somehow for the majority of triangle
types?  I'll see how that looks soon.